Things are starting to look good
The main things to implement now is logic for checking if the ping was
successful (atm it will always show "OK"), The CPU percentage thingy in
the header does not work as intended.
Otherwise, things look good, hosts are parsed-pinged-presented nicely,
fancy countdown.
Testing (like existence of settings.cfg) and general error handling is
also pending.