V 1.1
New feature/bug fix, so it's not V.1.0 anymore:
WD1337 now detects if the terminal changes size (e.g when resizing a
PuTTY window) and resets the terminal after a completed ping round if a
change in terminal columns or terminal lines is detected.
So now WD1337 won't look like a complete mess when you resize the
terminal while WD1337 is running.